Move handler code to controller
