+++ product-export/src/connectors/post-deploy.js
@@ -1,5 +1,5 @@
 import { createApiRoot } from '../clients/create.client.js';
-import { createChangedProductSubscription } from './actions.js';
+import { createPublishedProductSubscription } from './actions.js';
 
 const CONNECT_GCP_TOPIC_NAME_KEY = 'CONNECT_GCP_TOPIC_NAME';
 const CONNECT_GCP_PROJECT_ID_KEY = 'CONNECT_GCP_PROJECT_ID';
@@ -9,7 +9,7 @@ async function postDeploy(properties) {
   const projectId = properties.get(CONNECT_GCP_PROJECT_ID_KEY);
 
   const apiRoot = createApiRoot();
-  await createChangedProductSubscription(apiRoot, topicName, projectId);
+  await createPublishedProductSubscription(apiRoot, topicName, projectId);
 }
 
 async function run() {
