+++ product-export/src/constants/http.status.constants.js
@@ -1,4 +1,3 @@
 export const HTTP_STATUS_SUCCESS_ACCEPTED = 202;
-export const HTTP_STATUS_SUCCESS_NO_CONTENT = 204;
 export const HTTP_STATUS_BAD_REQUEST = 400;
 export const HTTP_STATUS_RESOURCE_NOT_FOUND = 404;
