+++ product-export/test/integration/product-published.controller.spec.js
@@ -1,7 +1,7 @@
 import { expect, describe, afterAll, it } from '@jest/globals';
 import request from 'supertest';
 import server from '../../src/index.js';
-import { HTTP_STATUS_BAD_REQUEST } from '../../src/constants/http.status.constants.js';
+import {HTTP_STATUS_BAD_REQUEST, HTTP_STATUS_RESOURCE_NOT_FOUND} from '../../src/constants/http.status.constants.js';
 
 /** Reminder : Please put mandatory environment variables in the settings of your github repository **/
 describe('Test product-published.import.controller.js', () => {
@@ -11,7 +11,7 @@ describe('Test product-published.import.controller.js', () => {
 
     response = await request(server).post(`/`);
     expect(response).toBeDefined();
-    expect(response.statusCode).toEqual(404);
+    expect(response.statusCode).toEqual(HTTP_STATUS_RESOURCE_NOT_FOUND);
   });
 
   it(`When payload body does not exist, it should returns 400 http status`, async () => {
